requirements.txt
pydantic~=2.5.3
datefinder~=0.7.3
aiogram~=3.3.0
APScheduler~=3.10.4
python-dotenv~=1.0.0
pydantic-settings~=2.1.0
motor~=3.3.2
redis~=5.0.1